% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{cor2cov}
\alias{cor2cov}
\alias{cov2cor}
\title{Conversions between covariance and correlation matrices}
\usage{
cor2cov(sds, cormat = diag(length(sds)), set.names = names(sds))

cov2cor(covmat, set.names = rownames(covmat))
}
\arguments{
\item{sds}{A vector of SDs}

\item{cormat}{A correlation matrix}

\item{set.names}{Override dimension names}

\item{covmat}{A covariance matrix}
}
\value{
A (named) covariance/correlation matrix
}
\description{
This function converts a correlation matrix into a covariance matrix, given a vector of standard deviations, and the reverse.
}
\section{Functions}{
\itemize{
\item \code{cov2cor}: Retrieving a correlation matrix
}}

