% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{default.random.cov}
\alias{default.random.cov}
\alias{default.fixed.means}
\title{Specifying effect sizes for simulations}
\usage{
default.random.cov(design, include = NULL, include.factors = NULL,
  random.intercepts = TRUE, random.slope.interactions = TRUE, sd = 1,
  contrasts = NULL)

default.fixed.means(design, mean = 0, interactions = TRUE,
  intercept = TRUE, contrasts = NULL)
}
\arguments{
\item{design}{A design object.}

\item{include}{A character vector of fixed effect to include as random slopes. If NULL, all possible slopes are considered.}

\item{include.factors}{Which factors to include, NULL for all}

\item{random.intercepts}{If TRUE, the variance matrices contain random intercepts.}

\item{random.slope.interactions}{If TRUE, variance matrices contain random slope interactions.}

\item{sd}{Default standard deviation to use for all random variances.}

\item{contrasts}{Contrasts to use in the fixed/random effects. This should match the contrasts passed to the model.}

\item{mean}{Default means to use for all fixed effects.}

\item{interactions}{If TRUE, a fixed slope for each contrast is included.}

\item{intercept}{If TRUE, a fixed intercept is included.}
}
\value{
A list of variance matrices (for default.random.cov) or a named numeric vector (for default.fixed.means).
}
\description{
For power simulations, you need to specify effect sizes for the contrasts in the design. These functions help you set up those effects in a format that is expected by the power simulation routines.
}
\section{Functions}{
\itemize{
\item \code{default.fixed.means}: Specify default fixed means
}}

