% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{design.formula}
\alias{design.formula}
\title{Preparing Designs for Model Analysis}
\usage{
design.formula(design, contrasts = NULL,
  expand.contrasts = !is.null(contrasts), interactions = T,
  intercepts = T, response = "response", env = parent.frame())
}
\arguments{
\item{design}{The `factor.design` to be used.}

\item{interactions}{Should fixed effects be additive or interactive?}

\item{lhs}{The left-hand side of the equation. Typically, this is just the response/dependent variable.}

\item{ranefs}{Should random factors be included as random effects for lme4?}

\item{fixed.interactions}{Overwrite `interactions` for fixed effects}

\item{random.interactions}{Overwrite `interactions` for random effects.}
}
\value{
A `formula` object
}
\description{
Based on a given design, this function creates a model formula that can be used to analyze data with functions such as lm or lmer.
}
