% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dim.factor.design}
\alias{dim.factor.design}
\title{Design matrix dimensions}
\usage{
nrow(design)
ncol(design)
dim(design)
nobs(design)
}
\arguments{
\item{design}{A factor design object.

You can use nrow(design) or nobs(design) to retrieve the number of rows (planned observations) and ncol(design) to retrieve the number of columns of the design. dim(design) retrieves both as a 2-length integer vector.}
}
\description{
Design matrix dimensions
}
