% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{extend.design.matrix}
\alias{extend.design.matrix}
\alias{extend.design}
\title{Extend an existing design matrix}
\usage{
extend.design.matrix(df, factor, n)

extend.design(design, factor, n)
}
\arguments{
\item{df}{A design matrix (as a data frame)}

\item{factor}{The name of the factor along which to extend the design}

\item{n}{The number of complete replications}

\item{design}{A design object}
}
\value{
An extended design matrix, n-times a long as the original matrix.
}
\description{
Extend the experimental design by replicating the codes along a given fixed or random factor.
}
\section{Functions}{
\itemize{
\item \code{extend.design}: Extend a design object
}}

