% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/methods.R
\docType{class}
\name{factorContainer-class}
\alias{factorContainer-class}
\alias{show,factorContainer-method}
\title{Design matrix S4 functions}
\usage{
\S4method{show}{factorContainer}(object)
}
\arguments{
\item{object}{A \code{factorDesign} object.}
}
\description{
Design matrix S4 functions
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Display a factor container
}}

\examples{
des <- factor.design()
des <- fixed.factor("Factor1", c("1A","1B")) +
       fixed.factor("Factor2", c("2A","2B")) +
       random.factor("Subject", c("Factor1"))

}
