% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{factorDesign-class}
\alias{factorDesign-class}
\title{S4 Methods for designFactor}
\arguments{
\item{x}{a \code{factorDesign}}
}
\description{
S4 Methods for designFactor
}
\examples{
x <- fixed.factor("Factor1", c("1A","1B"))
y <- random.factor("Subject", c("Factor1"))

}
