% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\docType{data}
\name{gibsonwu2013}
\alias{gibsonwu2013}
\title{Gibson & Wu (2013)}
\description{
The dataset \code{gibsonwu2013} contains data from self-paced reading in Chinese, comparing the processing of subject-extracted relative clauses (SRCs) and object-extracted relative clauses (ORCs) in supportive contexts.
}
\references{
Gibson, E., & Wu, H.-H. I. (2013). Processing Chinese relative clauses in context. \emph{Language and Cognitive Processes, 28}, 125–155. \doi{10.1080/01690965.2010.536656}
}
\keyword{data}
