% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{is.random.factor}
\alias{is.random.factor}
\alias{is.fixed.factor}
\alias{is.factor.design}
\alias{is.design.factor}
\title{Checking factor design data types}
\usage{
is.random.factor(fac)

is.fixed.factor(fac)

is.factor.design(fac)

is.design.factor(fac)
}
\arguments{
\item{fac}{Object to check.}
}
\value{
`TRUE` or `FALSE`
}
\description{
Check if argument is a design factor (either random or fixed factor), specifically a random factor, a fixed factor or a factor design.
}
\section{Functions}{
\itemize{
\item \code{is.random.factor}: Check if argument is a random factor.

\item \code{is.fixed.factor}: Check if argument is a fixed factor.

\item \code{is.factor.design}: Check if argument is a factor design.
}}

