% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{model.power.sim}
\alias{model.power.sim}
\title{Run a power simulation for a fitted linear (mixed) model or factorial design.}
\usage{
model.power.sim(model, tests, nsim = 1000L, cl = NULL, alpha = NULL)
}
\arguments{
\item{model}{A fitted model object}

\item{tests}{A list of tests to run for each simulation}

\item{nsim}{Number of iterations}

\item{cl}{A cluster object to parallelize the simulations}

\item{alpha}{The alpha level (Type-I error) to use for binary H0/H1 decisions}
}
\value{
A matrix with the results of all simulations
}
\description{
Run a power simulation for a fitted linear (mixed) model or factorial design.
}
\seealso{
[design.power.sim()], [tests()]
}
