% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{nobs,factorDesign-method}
\alias{nobs,factorDesign-method}
\title{Retrieve the number of observations}
\usage{
\S4method{nobs}{factorDesign}(object)
}
\arguments{
\item{object}{A designFactor object}
}
\value{
The number of observations
}
\description{
Retrieve the number of observations
}
