% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{output.design}
\alias{output.design}
\title{Summary of Factor Designs}
\usage{
output.design(design, group_by = NULL, order_by = NULL,
  randomize = FALSE, random = function(id, fac)
  factor(sprintf("\%s\%d", fac, id), sprintf("\%s\%d", fac,
  sort(unique(id)))))
}
\arguments{
\item{design}{The `factorDesign` object to summarize.}

\item{group_by}{If not `NULL`, the design matrix is grouped by these factors. Factors must be valid columns of the design matrix. If used, `$codes` will be a list matched to the entries in `$groups`.}

\item{order_by}{If not `NULL`, output within each output group is ordered by these columns.}

\item{randomize}{After ordering, remaining rows in the same order rank are randomly shuffled.}

\item{random}{A function used for naming random factor levels. Should be vectorized and accept the ID (integer) as a first argument and the name (single character value) of the random factor. Functions such as as.double() or as.integer() *are* possible because they ignore the second argument and only convert the ID.}
}
\value{
A list containing the output summary, including the following named entities:

   *$table*: Either a tibble with all experimental codes or a list of tibbles of experimental codes. The list entries are matched to the rows of `$groups`.

   *$groups*: If grouped, contains a tibble in which each row represents an output group, matched to the entries in $codes. If not grouped, this is `NULL`.

   *$ordered*: If ordered, contains a vector of order criteria. If not ordered, this is `NULL`.

   *$randomized*: Value of `randomized`.

   *$units*: A list of random factors and their levels for this design as tibbles. Empty list if no random factors in the design.

   *$formulas*: Example model formulae for use with functions such as lm() and lmer().
}
\description{
This function creates a useful summary of a factor design, including the design matrix itself as well as other parameters and a list of random factors as experimental units.
}
\seealso{
[design.formula()] for more options generating model formulae other than the suggested default ones in the summary.
}
