% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\docType{methods}
\name{plot,designr.power.curve-method}
\alias{plot,designr.power.curve-method}
\title{Run a power simulation for a factorial design.}
\usage{
\S4method{plot}{designr.power.curve}(x, ci = 0.95,
  ignore.failed.models = TRUE, add_p_breaks = c(attr(sim.out,
  "alpha_level"), 0.8), ncol = 3, plot.mean = "p", plot.error = "e",
  ...)
}
\arguments{
\item{x}{A power curve object (from design.power.curve(...))}

\item{ci}{The width of the confidence interval to plot}

\item{ignore.failed.models}{Should failed model fits be ignored (TRUE) or counted as failure to reject the null hypothesis (FALSE)}

\item{add_p_breaks}{Add horizontal dashed lines at those p-values (e.g., alpha from the power curve object and beta=.8)}

\item{ncol}{Number of plot facets in each horizontal line. Each facet is a different test.}

\item{plot.mean}{How should the mean power (beta) be plotted? Valid values are: line, l, point, p}

\item{plot.error}{How should the confidence interval be plotted? Valid values are: line, l, errorbar, e, ribbon, r, point, p}

\item{...}{(ignored)}
}
\value{
A ggplot plot
}
\description{
Run a power simulation for a factorial design.
}
\seealso{
[design.power.sim()], [design.power.curve()], [ggplot2::ggplot()]
}
