% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{+.factor.container}
\alias{+.factor.container}
\title{Concatenate design factors and designs}
\usage{
\method{+}{factor.container}(e1, e2)
}
\description{
By adding factors and designs by "+", a new design is created that contains all of the components.
}
