% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{+.factorContainer}
\alias{+.factorContainer}
\title{Concatenate design factors and designs}
\usage{
\method{+}{factorContainer}(e1, e2)
}
\description{
By adding factors and designs by "+", a new design is created that contains all of the components.
}
