% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.factor.container}
\alias{print.factor.container}
\alias{show.factor.container}
\title{Output a design factor summary}
\usage{
print.factor.container(x)

show.factor.container(object)
}
\description{
Output a design factor summary
}
\section{Functions}{
\itemize{
\item \code{show.factor.container}: Show a design factor summary
}}

