% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show.factorContainer}
\alias{show.factorContainer}
\title{Output a design factor summary}
\usage{
show.factorContainer(object)
}
\arguments{
\item{object}{the factor container to display}
}
\description{
Output a design factor summary
}
