% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simulate.lm}
\alias{simulate.lm}
\title{Simulate and fit a linear model based on the given design and effects.}
\usage{
\method{simulate}{lm}(design, formula = NULL,
  fixefs = default.fixed.means(design),
  varcov = default.random.cov(design, sd = 0), residual.sd = 1,
  collapse = names(random.factors(design))[-1], collapse.method = mean,
  contrasts = NULL, empirical = F)
}
\arguments{
\item{design}{A design object}

\item{formula}{The model formula to be fitted (using stats::lm)}

\item{fixefs}{A named vector of fixed effect sizes}

\item{varcov}{A list of covariance matrices, one per random factor. Names of list entries should match random factor names.}

\item{residual.sd}{The standard deviation of the residual term}

\item{collapse}{Random factors to collapse (by default, all but the first)}

\item{collapse.method}{Method to use to collapse observations (default: mean)}

\item{contrasts}{Named list of contrast matrices to override.}
}
\value{
A fitted lm object
}
\description{
Simulate and fit a linear model based on the given design and effects.
}
\seealso{
[lm()]
}
