% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simulate.response}
\alias{simulate.response}
\title{Simulate a response variable}
\usage{
simulate.response(design, contrasts = NULL,
  means = default.fixed.means(design, contrasts = contrasts),
  varcov = default.random.cov(design, contrasts = contrasts, include =
  names(means), sd = 0), residual.sd = 1, collapse.contrasts = T,
  empirical = F)
}
\arguments{
\item{design}{The factor design to use}

\item{contrasts}{Contrast codes to be used}

\item{means}{A named numeric vector with each mean (fixed effect)}

\item{residual.sd}{The residual error standard deviation in each cell of the design}

\item{empirical}{Are the means and (co-)variances empirical or population values? If true, analyzing the simulated response should yield the recovered parameters. If false, a sufficiently large sample of analyses should be representative of those values (appropriate for power simulations).}

\item{varvoc}{Variance-covariance matrices for each level. This should be a named list of named symmetric matrices. Names of list entries should match the names of the random factors and row/column names of the listed matrices should match the names of the effects (ideally the same as in `means`).}

\item{include}{Only use these factors (instead of the factors listed in `means`)}
}
\description{
Given a factor design, cell means (fixed effects) and random-level covariance matrices (random effects), simulate a normally distributed response variable.
}
