% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simulate_lmer}
\alias{simulate_lmer}
\title{Mixed model simulation}
\usage{
simulate_lmer(design, formula = NULL,
  fixefs = default.fixed.means(design),
  varcov = default.random.cov(design, sd = 1), residual.sd = 1,
  contrasts = NULL, empirical = FALSE)
}
\arguments{
\item{design}{A design object}

\item{formula}{The model formula to be fitted (using lme4::lmer)}

\item{fixefs}{A named vector of fixed effect sizes}

\item{varcov}{A list of covariance matrices, one per random factor. Names of list entries should match random factor names.}

\item{residual.sd}{The standard deviation of the residual term}

\item{contrasts}{Named list of contrast matrices to override.}

\item{empirical}{Are the means and (co-)variances empirical or population values?}
}
\value{
A fitted lmerMod object
}
\description{
Simulate and fit a linear mixed model based on the given design and effects.
}
\seealso{
[lme4::lmer()]
}
