% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simulate_power}
\alias{simulate_power}
\title{Run a power simulation for a factorial design.}
\usage{
simulate_power(design, formula = NULL, tests, modelClass = "lmer",
  contrasts = NULL, fixefs = default.fixed.means(design, contrasts =
  contrasts), varcov = default.random.cov(design, contrasts = contrasts,
  sd = if (modelClass == "lm" || identical(modelClass, stats::lm)) 0 else
  1), residual.sd = 1, ...)
}
\arguments{
\item{design}{A design object}

\item{formula}{The model formula to be fitted (using lme4::lmer or stats::lm)}

\item{tests}{A list of tests to run for each simulation}

\item{modelClass}{The model class to use (either "lm" or "lmer")}

\item{contrasts}{Named list of contrast matrices to override.}

\item{fixefs}{A named vector of fixed effect sizes}

\item{varcov}{A list of covariance matrices, one per random factor. Names of list entries should match random factor names.}

\item{residual.sd}{The standard deviation of the residual term}

\item{...}{More arguments to pass to model.power.sim, such as alpha, cl, or nsim}
}
\value{
A matrix with the results of all simulations
}
\description{
Run a power simulation for a factorial design.
}
\seealso{
[model.power.sim()], [tests()], [lme4::lmer()], [stats::lm()]
}
