% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simulate_power_curve}
\alias{simulate_power_curve}
\title{Run an array of mixed-effects power simulations for a factorial design with different sample sizes}
\usage{
simulate_power_curve(design, formula = NULL, contrasts = NULL, along,
  ns, tests, alpha = 0.05, ci = NULL, ...)
}
\arguments{
\item{design}{A design object}

\item{formula}{The model formula to be fitted (using lme4::lmer or stats::lm)}

\item{contrasts}{Named list of contrast matrices to override.}

\item{along}{The name of the random factor along which to extend the matrix and simulate power.}

\item{ns}{The replication factors (integer values, such as 1:4) to use to extend the design. Power is simulated for each replication separately.}

\item{tests}{A list of tests to run for each simulation}

\item{alpha}{The alpha level (Type-I error) to use for the simulations}

\item{ci}{The confidence interval for beta (Type-II error)}

\item{...}{More arguments to pass to design.power.sim, such as cl, nsim, fixefs ect.}
}
\value{
A matrix with the aggregated results of all power simulations
}
\description{
Run an array of mixed-effects power simulations for a factorial design with different sample sizes
}
\seealso{
[model.power.sim()], [design.power.sim()], [tests()], [lme4::lmer()], [stats::lm()]
}
