% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{[.factor.design}
\alias{[.factor.design}
\title{Retrieve factors or design matrix from an factor design object}
\usage{
design[i, col]
design[i, NULL]
design[, col]
design[, NULL]
design[[fac]]
design[facs]
}
\arguments{
\item{i}{The row number (number of the planned observation) to retrieve from the design matrix. Leave empty (but don't leave the comma out!) to select all rows.}

\item{col}{The column[s] (factor name[s]) to select from the matrix. Leave empty (but don't leave the comma out!) to select all columns.}

\item{fac, facs}{The factor name(s) to select from the factor list.

The design[i, col] syntax can be used to retrieve the design matrix (instead of the design object itself). You may leave names/indices for rows out in order to select all rows. You ALWAYS have to specify columns by names, indices or select all by using NULL or a single TRUE value.

If you use design[facs], you are selecting factor objects as a list from the design object, specified by names or indices in facs.

If you use design[[fac]], you are selecting a single factor object, specified by a name or index.}
}
\description{
Retrieve factors or design matrix from an factor design object
}
