% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{subset,factorDesign-method}
\alias{subset,factorDesign-method}
\alias{subset}
\title{Subset factor designs}
\usage{
subset(x, ...)
}
\arguments{
\item{x}{A factorDesign object}

\item{...}{*subset*: Criteria along which to filter in planned observations / design matrix rows., *select*: Names of columns to keep in the design matrix}
}
\description{
Subset factor designs
}
