% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subset.factorDesign}
\alias{subset.factorDesign}
\title{Subset factor designs}
\usage{
subset.factorDesign(x, subset, select = names(x), ...)
}
\arguments{
\item{x}{The factor design to subset}

\item{subset}{Criteria along which to filter in planned observations / design matrix rows.}

\item{select}{Names of columns to keep in the design matrix}

\item{...}{(ignored)}
}
\value{
A factor design with updated design matrix.
}
\description{
Subset factor designs
}
