% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{tests}
\alias{tests}
\title{Specify the list of tests to perform on each power simulation. Implemented tests are:}
\usage{
tests(...)
}
\arguments{
\item{...}{A list of tests}
}
\value{
A ggplot plot
}
\description{
t(x) : Use lmerTest to calculate the p-value for the factor x
}
\details{
t.(x) : Calculate p-value from t-values by approximating df=Inf

simr(type, x, y) : Use the simr package to perform test type("x", "y")
}
\examples{

# These are three different ways to specify a t-test on the same factor condition1
# Each of them is performed for each simulation
tests(
   t(condition1),
   t.(condition1),
   simr(fixed, condition1, t)
)

}
\seealso{
[design.power.sim()], [design.power.curve()], [model.power.sim()], [simr], [lmerTest]
}
