% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{design.formula}
\alias{design.formula}
\title{Preparing Designs for Model Analysis}
\usage{
design.formula(design, contrasts = NULL,
  expand.contrasts = !is.null(contrasts), interactions = T,
  intercepts = T, response = "dv", env = parent.frame())
}
\arguments{
\item{design}{The `factorDesign` to be used.}

\item{contrasts}{The contrasts to override (NULL if none to override)}

\item{expand.contrasts}{If TRUE, factors with more than one contrast are replaced by so many contrasts, i.e. the result contains the names of the individual contrasts, not of the factors.}

\item{interactions}{Should fixed effects be additive or interactive?}

\item{intercepts}{Should an intercept be included?}

\item{response}{The left-hand side of the equation. Typically, this is just the response/dependent variable.}

\item{env}{The environment in which to embed the formula}
}
\value{
A named list of `formula` objects, each name corresponding to the type of model it is suited for.
}
\description{
Based on a given design, this function creates a model formula that can be used to analyze data with functions such as lm or lmer.
}
