% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{nobs,factorDesign-method}
\alias{nobs,factorDesign-method}
\title{Retrieve the number of observations}
\usage{
\S4method{nobs}{factorDesign}(object)
}
\arguments{
\item{object}{A designFactor object}
}
\value{
The number of observations
}
\description{
Retrieve the number of observations
}
\examples{
des <- fixed.factor("Factor1", c("1A","1B")) +
       fixed.factor("Factor2", c("2A","2B")) +
       random.factor("Subject", c("Factor1"))
       
nobs(des)

stopifnot(nobs(des) == 4)

}
