% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{output.design}
\alias{output.design}
\title{Summary of Factor Designs}
\usage{
output.design(design, group_by = NULL, order_by = NULL,
  randomize = F)
}
\arguments{
\item{design}{The `factor.design` object to summarize.}

\item{group_by}{If not `NULL`, the design matrix is grouped by these factors. Factors must be valid columns of the design matrix. If used, `$codes` will be a list matched to the entries in `$groups`.}

\item{order_by}{If not `NULL`, output within each output group is ordered by these columns.}

\item{randomize}{After ordering, remaining rows in the same order rank are randomly shuffled.}
}
\value{
A list containing the output summary, including the following named entities:

   *$table*: Either a `data.frame` with all experimental codes or a list of data.frames of experimental codes. The list entries are matched to the rows of `$groups`.

   *$groups*: If grouped, contains a data.frame in which each row represents an output group, matched to the entries in $codes. If not grouped, this is `NULL`.

   *$ordered*: If ordered, contains a vector of order criteria. If not ordered, this is `NULL`.

   *$randomized*: Value of `randomized`.

   *$units*: A list of random factors and their levels for this design as data.frames. Empty list if no random factors in the design.

   *$fixed.model.formula*: An example model formula for use with functions such as lm().

   *$mixed.model.formula*: An example model formula for use with functions such as lmer().
}
\description{
This function creates a useful summary of a factor design, including the design matrix itself as well as other parameters and a list of random factors as experimental units.
}
\seealso{
[design.formula()] for more options generating model formulae other than the suggested default ones in the summary.
}
