% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{+,factorContainer,factorContainer-method}
\alias{+,factorContainer,factorContainer-method}
\title{Concatenate design factors and designs}
\usage{
\S4method{+}{factorContainer,factorContainer}(e1, e2)
}
\arguments{
\item{e1, e2}{factor containers, such as factors or designs}
}
\value{
A factorDesign object
}
\description{
By adding factors and designs by \code{+}, a new design is created that contains all of the components.
}
\examples{

des <- fixed.factor("Factor1", c("1A","1B")) +
       fixed.factor("Factor2", c("2A","2B")) +
       random.factor("Subject", c("Factor1"))


}
